[qtAliceVision] SingleImageLoader: Catch errors when loading image
If an exception is caught when attempting to load an image, instead
of just logging the error message and keeping on going, a flag is set
in the `ResponseData` object.

The flag's value depends on the type of exception/error that is caught.
For now, there are only two types of errors: either the file cannot be found
(`MISSING_FILE`) or cannot be opened (`ERROR`).